Chore: moved MixerProtocol Interfaces into seperate file: MixerProtocolInterface
