+++ src/constants/mixerProtocols/EmberLawo.ts
@@ -1,4 +1,4 @@
-import { IMixerProtocol, emptyMixerMessage } from '../MixerProtocolPresets';
+import { IMixerProtocol, emptyMixerMessage } from '../MixerProtocolInterface';
 import path from 'path';
 import os from 'os';
 import fs from 'fs';
