+++ src/constants/mixerProtocols/midasMaster.ts
@@ -1,4 +1,4 @@
-import { IMixerProtocol, emptyMixerMessage } from '../MixerProtocolPresets';
+import { IMixerProtocol, emptyMixerMessage } from '../MixerProtocolInterface';
 
 export const MidasMaster: IMixerProtocol = {
     protocol: 'OSC',
